app_lvl1/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pypdf
python-dotenv